==> analyses/Main_Analysis/04-create.variation.plots.qmd <==
---
format:
    html:
        embed-rmd: true
        output-ext: "htm"
---


```{r, eval=T, echo=F, results=FALSE, message = FALSE, warning = FALSE}
if (file.exists("renv.lock")){
    renv::restore(prompt = FALSE)
}

FN <- "../../../../workdir/temp/temp.workspace.RData"
if (file.exists(FN)){
  load(FN)
} else {
  exit("No workspace found. Please run the previous step first and make sure it finished successfully.")
}

tabVar <- ".tabset .tabset-fade .tabset-pills"

if (createNew){
  cat("Running in resume mode - data for this module is only available if run on a full analysis.")
}

```

```{r, eval=T, echo=F, results=FALSE, message = FALSE, warning = FALSE}

```

<!-- Set PArameters Module -->
<!-- Set the chnkPrefix to make all chunks unique in the overall folder -->
```{r partB_create_variation_plots, echo=createNew, eval=createNew, warning=FALSE}
chnkPrefix <- "partB.integrated.seurat.object."
VersionPdfExt <- VersionPdfExt <- paste0(".",chnkPrefix,"V", gsub("-", "", Sys.Date()), ".pdf")

```



```{r partB_create-individual-varFeatures, echo=createNew, eval=createNew, warning=FALSE, results=F}
###############################################################################
## Do percent mt plots                                                       ##
plotList <- list()
chnkVec <- as.vector(NULL, mode = "character")

## First make variation plot for integrated samples, than for all individual samples separately
tag <- "Integrated_Samples"
Seurat::DefaultAssay(OsC) <- "RNA"


OsC <- Seurat::FindVariableFeatures(
    object = OsC,
    selection.method = 'vst', 
    nfeatures = Obio@parameterList$NtopGenes
)

# Identify the 10 most highly variable genes
label2000 <- paste0("integrated", "_", "top2000var")
Obio@dataTableList$referenceList[[ label2000]]<- head(
    x = Seurat::VariableFeatures(object = OsC), 
    2000
)
    
label30 <- paste0("integrated", "_", "top30var")
    Obio@dataTableList$referenceList[[ label30]]<- head(
    x = Seurat::VariableFeatures(object = OsC), 
    30
)


## slot for variable features OsC@assays$RNA@var.features

dfVar <- OsC@assays$RNA@meta.data
names(dfVar) <- gsub("vf_", "",names(dfVar))
dfVar$var.features <- NULL
dfVar$var.feature.rank <- NULL
dfVar[["gene"]] <- row.names(OsC)

names(dfVar) <-  gsub("vst_counts_", "", names(dfVar))

## Workaraound to get data frame
OsC@meta.data[["all"]] <- "all"
OsC@meta.data[1, "all"] <- "fix"
Seurat::Idents(OsC) <- "all"
    
cluster.averages <- Seurat::AggregateExpression(
    OsC, 
    return.seurat = TRUE
)
    
Seurat::Idents(OsC) <- "sampleID"
    
dfAvgExpr <- data.frame(cluster.averages@assays$RNA$data)
dfAvgExpr$fix <- NULL
dfAvgExpr[["gene"]] <- row.names(dfAvgExpr)
names(dfAvgExpr)[1] <- "Avg.Expression"

dfVar <- merge(
    dfVar, 
    dfAvgExpr, 
    by.x = "gene",
    by.y = "gene"
)

dfVar[["Type"]] <- "Standard"
dfVar[dfVar$gene %in% Obio@dataTableList$referenceList[[ label2000]], "Type"] <- "Most Variable"

dfVar[["text"]] <- ""

dfVar[dfVar$gene %in% as.vector(Obio@dataTableList$referenceList[[label30]]), "text"] <- dfVar[dfVar$gene %in% as.vector(Obio@dataTableList$referenceList[[label30]]), "gene"]

#dotsize <- 0.5

dotsize  = 1


if (nrow(OsC@meta.data) > 50000){
  dotsize  = 0.05
} else if (nrow(OsC@meta.data) > 20000){
  dotsize = 0.1
} else if (nrow(OsC@meta.data) > 10000){
  dotsize = 0.5
} else if (nrow(OsC@meta.data) > 5000){
  dotsize = 0.75
}

legendDotSize <- Obio@parameterList$legendDotSize

plotList[[tag]] <- ggplot2::ggplot(
    data = dfVar, 
        ggplot2::aes(
            x=Avg.Expression, 
            y=variance.standardized, label = text, color = Type
        )
  ) + ggrastr::rasterize(ggplot2::geom_point( shape=16, size = dotsize), dpi = rasterDpi
  ) + ggplot2::xlab("Average Expression") + ggplot2::ylab("Variance Standarized"
  ) + ggplot2::theme_bw(
  ) +  ggplot2::theme(
    axis.text.y   = ggplot2::element_text(size=8),
    axis.text.x   = ggplot2::element_text(size=8),
    axis.title.y  = ggplot2::element_text(size=8),
    axis.title.x  = ggplot2::element_text(size=8),
    axis.line = ggplot2::element_line(colour = "black"),
    panel.border = ggplot2::element_rect(colour = "black", fill=NA, size=1),
    plot.title = ggplot2::element_text(hjust = 0.5, size = 12)
) + ggplot2::guides(col = ggplot2::guide_legend(override.aes = list(shape = 16, size = legendDotSize))    
) + ggplot2::ggtitle(paste0("Variance vs. Expression in the Overall Experiment")
) + ggplot2::scale_color_manual(values=c("#FF0000", "#000000")
) + ggrepel::geom_text_repel(
) 


    
###########################################################################
## Save plot to file                                                     ##
FNbase <- paste0("variation.integrated.samples", VersionPdfExt)
FN <- paste0(Obio@parameterList$reportFigDir, FNbase)
FNrel <- paste0("report_figures/", FNbase)
    
pdf(FN)
    print(plotList[[tag]])
dev.off()
##                                                                       ##
###########################################################################
    
if (exists("urlString") & ! is.null(urlString)){
    link <- paste0(
      'An interactive version of this figure can be found ',
      '<a href="https://',urlString,'/',Obio@parameterList$project_id,'/scatterplot?x_axis=',paste0(tag, '_AvgExpr'),'&y_axis=',paste0(tag, '_var_std" target = "_blank">here</a>.'))
}    

figCap <- paste0(
    '**Figure ',
    figureCount,
    ':** Variance versus averaged gene expression for overall sample.', 
    '. ',
    'Download a pdf of this figure <a href="',FNrel,'" target="_blank">here</a>. ',
     link
)
    
    
   
    
NewChnk <- paste0(
    "#### ",tag,
    "\n```{r varplot_",tag,", results='asis', echo=F, eval=TRUE, warning=FALSE, fig.cap='",figCap,"'}\n",
    "\n",
    "\n print(plotList[['",tag,"']])",
    "\n cat(  '\n')",
    "\n\n\n```\n"   
)
     
## Histogram Part C done                                                 ##
###########################################################################
    
       
chnkVec <- c(
    chnkVec,
    NewChnk
)
    
figureCount <- figureCount + 1
print(" All variation done.")

## Now the individual samples ##

xmax <- ceiling(max(dfVar$Avg.Expression))
ymax <- ceiling(max(dfVar$variance.standardized))

dfVarRes <- unique(
    dfVar[,c("gene", "Avg.Expression", "variance.standardized")]
)

names(dfVarRes) <- gsub(
    "Avg.Expression", paste0(tag, "_AvgExpr"), names(dfVarRes)
)

names(dfVarRes) <- gsub(
    "variance.standardized", paste0(tag, "_var_std"), names(dfVarRes)
)

## Make variation plot for individual samples ##

## Create sample list from Seurat object

SampleList <- Seurat::SplitObject(OsC, split.by = "sampleName")

if (length(SampleList) > 1){
for (i in 1:length(SampleList)){
    tag <- paste0("Ind_var_", names(SampleList)[i])
    
    
    Seurat::DefaultAssay(SampleList[[i]]) <- "RNA"

    SampleList[[i]] <- Seurat::FindVariableFeatures(object = SampleList[[i]])

    SampleList[[i]] <- Seurat::FindVariableFeatures(
        object = SampleList[[i]],
        selection.method = 'vst', 
        nfeatures = 2000
    )

    # Identify the 10 most highly variable genes
    label2000 <- paste0(names(SampleList)[i], "_", "top2000var")
    Obio@dataTableList$referenceList[[ label2000]]<- head(
        x = Seurat::VariableFeatures(object = SampleList[[i]]), 
        2000
    )
    
    label30 <- paste0(names(SampleList)[i], "_", "top30var")
    Obio@dataTableList$referenceList[[ label30]]<- head(
        x = Seurat::VariableFeatures(object = SampleList[[i]]), 
        30
    )


    ## slot for variable features OsC@assays$RNA@var.features

    dfVar <- SampleList[[i]]@assays$RNA@meta.data
    names(dfVar) <- gsub("vf_", "",names(dfVar))
    dfVar$var.features <- NULL
    dfVar$var.features.rank <- NULL
    dfVar[["gene"]] <- row.names(OsC)

    names(dfVar) <-  gsub("vst_counts_", "", names(dfVar))

    Seurat::Idents(SampleList[[i]]) <- "sampleName"

    ## Workaraound to get data frame
    SampleList[[i]]@meta.data[["all"]] <- "all"
    SampleList[[i]]@meta.data[1, "all"] <- "fix"
    Seurat::Idents(SampleList[[i]]) <- "all"

    cluster.averages <- Seurat::AggregateExpression(
        SampleList[[i]],
        return.seurat = TRUE
    )


    dfAvgExpr <- data.frame(cluster.averages@assays$RNA$data)
    dfAvgExpr$fix <- NULL
    dfAvgExpr[["gene"]] <- row.names(dfAvgExpr)
    names(dfAvgExpr)[1] <- "Avg.Expression"



    dfVar <- merge(
        dfVar, 
        dfAvgExpr, 
        by.x = "gene",
        by.y = "gene"
    )

    dfVar[["Type"]] <- "Standard"
    dfVar[dfVar$gene %in% Obio@dataTableList$referenceList[[ label2000]], "Type"] <- "Most Variable"

    dfVar[["text"]] <- ""

    dfVar[dfVar$gene %in% as.vector(Obio@dataTableList$referenceList[[label30]]), "text"] <- dfVar[dfVar$gene %in% as.vector(Obio@dataTableList$referenceList[[label30]]), "gene"]

    # dotsize <- 0.5

    

    plotList[[tag]] <- ggplot2::ggplot(
        data = dfVar, 
        ggplot2::aes(
            x=Avg.Expression, 
            y=variance.standardized, label = text, color = Type)
            ) + ggrastr::rasterize(ggplot2::geom_point( shape=16, size = dotsize ), dpi = rasterDpi ) + ggplot2::xlab("Average Expression") + ggplot2::ylab("Variance Standarized"
            ) + ggplot2::theme_bw(
            ) +  ggplot2::theme(
                axis.text.y   = ggplot2::element_text(size=8),
                axis.text.x   = ggplot2::element_text(size=8),
                axis.title.y  = ggplot2::element_text(size=8),
                axis.title.x  = ggplot2::element_text(size=8),
                axis.line = ggplot2::element_line(colour = "black"),
                panel.border = ggplot2::element_rect(colour = "black", fill=NA, size=1),
                plot.title = ggplot2::element_text(hjust = 0.5, size = 12)
            ) + ggplot2::guides(col = ggplot2::guide_legend(override.aes = list(shape = 16, size = legendDotSize))
            ) + ggplot2::ggtitle(paste0("Individual variance in the ", names(SampleList)[i],  " sample.")
            ) + ggplot2::scale_color_manual(values=c("#FF0000", "#000000")
            ) + ggrepel::geom_text_repel(
            ) +  ggplot2::xlim(0, xmax) + ggplot2::ylim(0, ymax
            ) 


    
    ###########################################################################
    ## Save plot to file                                                     ##
    FNbase <- paste0("Individual.var.features",names(SampleList)[i], VersionPdfExt)
    FN <- paste0(Obio@parameterList$reportFigDir, FNbase)
    FNrel <- paste0("report_figures/", FNbase)
    
    pdf(FN)
        print(plotList[[tag]])
    dev.off()
    ##                                                                       ##
    ###########################################################################
    
    if (exists("urlString") & ! is.null(urlString)){
        link <- paste0(
            'An interactive version of this figure can be found ',
            '<a href="https://',urlString,'/',Obio@parameterList$project_id,'/scatterplot?x_axis=',paste0(tag, '_AvgExpr'),'&y_axis=',paste0(tag, '_var_std" target="_blank">here</a>'),
            '. '
          )
    }
    

    figCap <- paste0(
        '**Figure ',
        figureCount,
        ':** Variance versus averaged gene expression for sample ', 
        names(SampleList)[i],
        '. ',
        'Download a pdf of this figure <a href="',FNrel,'" target="_blank">here</a>. ',
         link
    )
    
    NewChnk <- paste0(
        "#### ",tag,
        "\n```{r varplot_",tag,", results='asis', echo=F, eval=TRUE, warning=FALSE, fig.cap='",figCap,"'}\n",
        "\n",
        "\n print(plotList[['",tag,"']])",
        "\n cat(  '\n')",
        "\n\n\n```\n"   
    )
     
    ## Histogram Part C done                                                 ##
    ###########################################################################
    
       
    chnkVec <- c(
        chnkVec,
        NewChnk
    )
    
    figureCount <- figureCount + 1
    print(paste0(names(SampleList)[i], " individual variation done."))
    
    ## Add to result table ##
    dfVarTemp <- unique(
    dfVar[,c("gene", "Avg.Expression", "variance.standardized")]
    )
    
    names(dfVarTemp) <- gsub(
        "Avg.Expression", paste0(tag, "_AvgExpr"), names(dfVarTemp)
    )
    
    names(dfVarTemp) <- gsub(
        "variance.standardized", paste0(tag, "_var_std"), names(dfVarTemp)
    )
    
    dfVarRes <- merge(
        dfVarRes, 
        dfVarTemp, 
        by.x = "gene",
        by.y = "gene",
        all = TRUE
    )
    dfVarRes[is.na(dfVarRes)] <- 0
    
}
} # end if multiple samples

## Make sure summary goes first ##
Obio@dataTableList[["dfVariation"]] <- data.frame(NULL)
Obio@dataTableList[["dfVariation"]] <- dfVarRes

if (length(plotList) > 3){
    tabVar <- ".tabset .tabset-fade .tabset-dropdown"
} else {
    tabVar <- ".tabset .tabset-fade .tabset-pills"
}

## Done create cellRanger QC plots                                           ##
###############################################################################


```


## Variation Figures
### Quality Control Plots: Variable Features in the combined and all individual dataset {`r tabVar`}
:::: {.panel-tabset .nav-pills}
```{r create-var-feature-plot-II, echo=createNew, eval=createNew, warning=FALSE, results="asis"}
## plot list will be integrated in full figure ##
cat(paste(knitr::knit(text = chnkVec, quiet = T), collapse = '\n'))
```
::::

```{r, eval=createNew, echo=F, results=FALSE, message = FALSE, warning = FALSE}
tempDir <- "../../../../workdir/temp/"

if (!dir.exists(tempDir)){
  dir.create(tempDir, recursive = T)
}

if (exists("whiteListWorkspace")){
    rm(list = setdiff(ls(), whiteListWorkspace))
}

FN <- "../../../../workdir/temp/temp.workspace.RData"
save.image(FN)
```
