==> analyses/Main_Analysis/14A-heatmaps.qmd <==
---
format:
    html:
        embed-rmd: true
        output-ext: "htm"
---

## Heatmap Chapter
```{r, eval=T, echo=F, results=FALSE, message = FALSE, warning = FALSE}
if (file.exists("renv.lock")){
    renv::restore(prompt = FALSE)
}

FN <- "../../../../workdir/temp/temp.workspace.RData"
if (file.exists(FN)){
  load(FN)
} else {
  exit("No workspace found. Please run the previous step first and make sure it finished successfully.")
}

library(dplyr)
```

<!-- Set Parameters for this module -->
<!-- Set the chnkPrefix to make all chunks unique in the overall folder -->
```{r heatmaps_14A, echo=TRUE, eval=TRUE, warning=FALSE}
chnkPrefix <- "partB.heatmaps.14A."
VersionPdfExt <- VersionPdfExt <- paste0(
  ".", chnkPrefix,"V", gsub("-", "", Sys.Date()), ".pdf"
)

if (!requireNamespace("ComplexHeatmap")){
  renv::install("jokergoo/ComplexHeatmap")
}


```



```{r select_granularity, echo=TRUE, eval=TRUE, warning=FALSE, results=F}

###############################################################################
## Depending on the project, another parameter can be used for heatmap       ##
## segmentation. The default is cluster_sampleName

segmentationCol <- "sampleName"

## Add column for averaging of cells
selClust <- paste0("clusterName_", segmentationCol)
OsC_HM <- OsC
OsC_HM@meta.data[[selClust]] <- paste0(
  OsC_HM@meta.data$clusterName,
  "_",
  OsC_HM@meta.data[,segmentationCol]
)



###############################################################################
## Calculataing average across segmentation                                  ##
Seurat::Idents(OsC_HM) <- selClust
OsCsel <- Seurat::AverageExpression(
  OsC_HM, 
  return.seurat = TRUE
)


if (segmentationCol == "sampleName" & length(grep("^sampleColor$", names(OsC_HM@meta.data)) == 1)){
  segmentationColorCol <- "sampleColor"
} else{
  segmentationColorCol <- paste0(
    segmentationCol,
    "Color"
  )
  
  ## Colors can also be set manually here
  allCols <- unique(OsC_HM@meta.data[,segmentationCol])
  colorSel <- scales::hue_pal()(length(allCols))
  
  OsC_HM@meta.data[[segmentationColorCol]] <- OsC_HM@meta.data[[segmentationCol]]
  for (i in 1:length(allCols)){
    OsC_HM@meta.data[[segmentationColorCol]] <- gsub(
      allCols[i],
      colorSel[i],
      OsC_HM@meta.data[[segmentationColorCol]]
      
    )
  }
  
} # done with else of if clause



selVec <- c(
  "clusterName", 
  "clusterColor",  
  segmentationCol, 
  segmentationColorCol, 
  selClust
)

dfAdd <- unique(OsC_HM@meta.data[,selVec])

## Add color vector for segmentationCol ##

#dfAdd[["treatmentGroupColor"]] <- dfAdd$meta_Treatment
#dfAdd$treatmentGroupColor <- gsub("Control", "#3B5EF7", dfAdd$treatmentGroupColor)
#dfAdd$treatmentGroupColor <- gsub("MatB52", "#4DA868", dfAdd$treatmentGroupColor)
#dfAdd$treatmentGroupColor <- gsub("Regeneron", "#EE6C58", dfAdd$treatmentGroupColor)

row.names(dfAdd) <- gsub("_", "-", dfAdd[,selClust])

OsCsel <- biologicToolsSC::addDf2seuratMetaData(
  obj = OsCsel, 
  dfAdd = dfAdd
)

## Order of factors
segmentationColOrder <- sort(unique(OsCsel@meta.data[[segmentationCol]]))

if (segmentationCol == "sampleName" & !is.null(Obio@parameterList$sampleNameOrder)){
  segmentationColOrder <- Obio@parameterList$sampleNameOrder
}

OsCsel@meta.data[[segmentationCol]] <- factor(
  OsCsel@meta.data[[segmentationCol]],
  levels = segmentationColOrder
)

OsCsel@meta.data$clusterName <- factor(
  OsCsel@meta.data$clusterName,
  levels=Obio@parameterList$clusterNameOrder
)


OsCsel@meta.data <- OsCsel@meta.data[order(OsCsel@meta.data$clusterName,OsCsel@meta.data[[segmentationCol]]),]


dfData <- data.frame(OsCsel@assays$RNA$data)
names(dfData) <- gsub("\\.", "-", names(dfData))

dataMatrix <- data.matrix(dfData)
dfMeta <- OsCsel@meta.data
 
dataMatrix <- dataMatrix[,gsub("_", "-", dfMeta[,selClust])]



###############################################################################
## Do percent mt plots  

HMgeneSelections <- list()

## Cluster markers

## Start with LR
LRgeneList <- list()
topNselection <- 3

dfTemp <-  Obio@dataTableList$dfClusterResiduals

if (is.null(Obio@parameterList$clusterNameOrder)){
    Obio@parameterList$clusterNameOrder <- sort(unique(OsC@meta.data$clusterName))
}


clusterNames <- unique( Obio@parameterList$clusterNameOrder)

for (i in 1:length(clusterNames)){
  dfTemp <- dfTemp[order(dfTemp[, clusterNames[i]], decreasing = T), ]
  
  if (topNselection > nrow(dfTemp)){
      topNselection <- nrow(dfTemp)
  }
  
  geneVec <- as.vector(sort(dfTemp[1:topNselection, "gene"]))
  
  LRgeneList[[clusterNames[i]]] <- geneVec
}

HMgeneSelections[["LR_Markers"]] <- LRgeneList

## Add LR_TF markers

## Get transcription factor genes ##
# if (dir.exists("/Volumes/babs/working/boeings/")){
#     hpc.mount <- "/Volumes/babs/working/boeings/"
# } else if (dir.exists("Y:/working/boeings/")){
#     hpc.mount <- "Y:/working/boeings/"
# } else if (dir.exists("/camp/stp/babs/working/boeings/")){
#     hpc.mount <- "/nemo/stp/babs/working/boeings/"
# } else {
#     hpc.mount <- ""
# }
#
#
# FN <- paste0(hpc.mount, "Projects/reference_data/pwd_folder/babs.txt")
#     dbTable <- read.delim(
#       FN,
#       header = F,
#       sep = "\t",
#       stringsAsFactors = F
#     )
# db.pwd <- as.vector(dbTable[1,1])

if (Obio@parameterList$geneIDcolumn != "mgi_symbol" & Obio@parameterList$geneIDcolumn != "hgnc_symbol") {
    queryGS <- "hgnc_symbol" 
} else {
    queryGS <- Obio@parameterList$geneIDcolumn
}


tempVec <- biologicSeqTools2::retrieve.gene.category.from.db(
    cat_id = "ag_lab_categories__10",
    password = db.pwd,
    gene.symbol = queryGS,
    user = Obio@parameterList$db.user,
    host = Obio@parameterList$host
)

dfTemp <-  Obio@dataTableList$dfClusterResiduals
dfTemp <- dfTemp[dfTemp$gene %in% tempVec, ]


clusterNames <- unique( Obio@parameterList$clusterNameOrder)

for (i in 1:length(clusterNames)){
  dfTemp <- dfTemp[order(dfTemp[, clusterNames[i]], decreasing = T), ]
  
  if (topNselection > nrow(dfTemp)){
      topNselection <- nrow(dfTemp)
  }
  
  geneVec <- as.vector(sort(dfTemp[1:topNselection, "gene"]))
  
  LRgeneList[[clusterNames[i]]] <- geneVec
}

HMgeneSelections[["LR_TF_Markers"]] <- LRgeneList


## Add DGE markers ##



## Make general markers heatmpa ##
dfGeneralMarkers <- Obio@dataTableList$dfGeneralMarkers 
dfGeneralMarkersPos <- dfGeneralMarkers[dfGeneralMarkers$avg_diff > 0,]


dfT <- unique(OsC@meta.data[,c("clusterName", "seurat_clusters")])
dfT[["DGEnames"]] <- paste0("C",dfT$seurat_clusters)

dfGeneralMarkersPos <- merge(
  dfGeneralMarkersPos,
  dfT,
  by.x = "cluster",
  by.y = "DGEnames"
)

dfTemp <- dfGeneralMarkersPos

DGEgeneList <- list()
topNselection <- 5


clusterNames <- unique( Obio@parameterList$clusterNameOrder)

for (i in 1:length(clusterNames)){
  dfTemp2 <- dfTemp[dfTemp$clusterName == clusterNames[i],]
  dfTemp2 <- dfTemp2[order(dfTemp2[, "avg_diff"], decreasing = T), ]
  
  if (topNselection > nrow(dfTemp2)){
      selVal <- nrow(dfTemp2)
  } else {
      selVal <- topNselection
  }
  
  geneVec <- as.vector(sort(dfTemp2[1:selVal, "gene"]))
  if (length(geneVec) > 0){
    DGEgeneList[[clusterNames[i]]] <- geneVec
  }
}


HMgeneSelections[["DGE_Markers"]] <- DGEgeneList

## DGE TF markers
## Make general markers heatmpa ##
dfGeneralMarkers <- Obio@dataTableList$dfGeneralMarkers 
dfGeneralMarkersPos <- dfGeneralMarkers[dfGeneralMarkers$avg_diff > 0,]


dfT <- unique(OsC@meta.data[,c("clusterName", "seurat_clusters")])
dfT[["DGEnames"]] <- paste0("C",dfT$seurat_clusters)

dfGeneralMarkersPos <- merge(
  dfGeneralMarkersPos,
  dfT,
  by.x = "cluster",
  by.y = "DGEnames"
)

dfTemp <- dfGeneralMarkersPos

## Get transcription factor genes ##
# if (dir.exists("/Volumes/babs/working/boeings/")){
#     hpc.mount <- "/Volumes/babs/working/boeings/"
# } else if (dir.exists("Y:/working/boeings/")){
#     hpc.mount <- "Y:/working/boeings/"
# } else if (dir.exists("/camp/stp/babs/working/boeings/")){
#     hpc.mount <- "/nemo/stp/babs/working/boeings/"
# } else {
#     hpc.mount <- ""
# }
#
#
# FN <- paste0(hpc.mount, "Projects/reference_data/pwd_folder/babs.txt")
#     dbTable <- read.delim(
#       FN,
#       header = F,
#       sep = "\t",
#       stringsAsFactors = F
#     )
# db.pwd <- as.vector(dbTable[1,1])

if (Obio@parameterList$geneIDcolumn != "mgi_symbol" & Obio@parameterList$geneIDcolumn != "hgnc_symbol") {
    queryGS <- "hgnc_symbol" 
} else {
    queryGS <- Obio@parameterList$geneIDcolumn
}


tempVec <- biologicSeqTools2::retrieve.gene.category.from.db(
    cat_id = "ag_lab_categories__10",
    password = db.pwd,
    gene.symbol = queryGS,
    user = Obio@parameterList$db.user,
    host = Obio@parameterList$host
)

dfTemp <- dfTemp[dfTemp$gene %in% tempVec, ]


DGEgeneList <- list()
topNselection <- 5


clusterNames <- unique( Obio@parameterList$clusterNameOrder)

for (i in 1:length(clusterNames)){
  dfTemp2 <- dfTemp[dfTemp$clusterName == clusterNames[i],]
  dfTemp2 <- dfTemp2[order(dfTemp2[, "avg_diff"], decreasing = T), ]
  
  if (topNselection > nrow(dfTemp2)){
      selVal <- nrow(dfTemp2)
  } else {
      selVal <- topNselection
  }
  
  geneVec <- as.vector(sort(dfTemp2[1:selVal, "gene"]))
  if (length(geneVec) > 0){
    DGEgeneList[[clusterNames[i]]] <- geneVec
  }
}

HMgeneSelections[["DGE_TF_Markers"]] <- DGEgeneList


## Load gene sets from cat reference file
catFN <-  Obio@parameterList$catRefFile

if (!is.null(catFN) && file.exists(catFN)){
  dfCat <- read.delim(
    catFN, 
    sep = "\t",
    stringsAsFactors = F
  )
}

## Add each cat to ref list 
for (i in 1:ncol(dfCat)){
  catName <- colnames(dfCat)[i]
  geneVec <- unique(
    as.vector(
      dfCat[3:nrow(dfCat), i]
    )
  )
  geneVec <- geneVec[geneVec %in% row.names(OsC)]
  catList <- list()
  catList[[catName]] <- geneVec
  HMgeneSelections[[catName]] <- catList
}


# ## Custom gene lists
# HMgeneSelections[["Custom_Top2_Markers"]]  <- list(
#     "GC1" = c("Galntl6", "Esrrg"),
#     "GC2" = c("Oxr1", "Adamts19"),
#     "GC3" = c("Hcn1", "Nrg3"),
#     "GC4" = c("Apoe", "Ptn"),
#     "GC5" = c("Dcc", "Mgat4c"),
#     "GC6" = c("Lrp1b", "Thsd7b"),
#     "GC7" = c("Macrod2", "Cntn5"),
#     "GC8" = c("Il1rapl2", "Kcnt2"),
#     "GC9" = c("Tmem132d", "Unc5d")
# )
# 
# HMgeneSelections[["Heatmap_Channels"]]  <- list(
#     "VGSC" = c("Scn1a", "Scn2a", "Scn1a","Scn2a", "Scn3a", "Scn4a","Scn5a", "Scn8a", "Scn9a", "Scn10a", "Scn11a"),
#     "VGPC" = c("Kcna2", "Kcna3", "Kcna4", "Kcna5", "Kcnb1", "Kcnb2", "Kcnc1", "Kcnc2", "Kcnc3", "Kcnc4", "Kcnd1", "Kcnd2","Kcnd3", "Kcnq1", "Kcnq2", "Kcnq3", "Kcnq4", "Kcnq5"),
#     "CaGC" = c("Cacna1a", "Cacna1b", "Cacna1c", "Cacna1d", "Cacna1e", "Cacna1f"),
#     "ClC" = c(
#       "Clcn1",
#       "Clcn2",
#       "Clcn3",
#       "Clcn4",
#       "Clcn5",
#       "Clcn6",
#       "Clcn7",
#       "Gabra1",
#       "Gabra2",
#       "Gabra3",
#       "Gabra4",
#       "Gabra5",
#       "Gabra6",
#       "Gabrb1",
#       "Gabrb2",
#       "Gabrb3",
#       "Gabrb4",
#       "Gabrb5",
#       "Gabrg1"
#     ),
#     "VGSCbeta" = c("Scn1b", "Scn2b", "Scn3b", "Scn4b"),
#     "IRPC" = c("Kcnj1", "Kcnj2","Kcnj3", "Kcnj4", "Kcnj5", "Kcnj6", "Kcnj8", "Kcnj9", "Kcnj10", "Kcnj11", "Kcnj12", "Kcnj13", "Kcnj14","Kcnj15", "Kcnj16","Kcnj18", "Kcnj19"),
#     "HCNC" = c("Hcn1", "Hcn2", "Hcn3", "Hcn4"),
#     "NaKExc" = c("Slc9a1", "Slc9a2", "Slc9a3", "Slc9a4","Slc9a5", "Slc9a6", "Slc9a7", "Slc9a8"),
#     "CaSenKC" = c("Kcnma1", "Kcnma2","Kcnma3","Kcnma4", "Kcnmb1", "Kcnmb2", "Kcnmb3", "Kcnmb4"),
#     "TrpC" = c("Trpa1","Trpa3", "Trpc1", "Trpc2","Trpc3", "Trpc4", "Trpc5", "Trpc6", "Trpc7", "Trpm1", "Trpm2","Trpm3","Trpm4", "Trpm5","Trpm6", "Trpm7","Trpm8","Trpv1","Trpv2","Trpv3", "Trpv4","Trpv5", "Trpv6")
# )


## done
################################



###############################################################################
## Function to create heatmaps                                               ##

addHeatmap2List <- function(
    dataMatrix,
    geneGroupList,
    cluster_column_slices   = FALSE,
    cluster_columns         = TRUE,
    relativeExpression = FALSE,
    cmdVec = NULL,
    tag = NULL,
    showAllLegends = TRUE,
    figureCount = 1,
    chnkVec = NULL, 
    plotList = NULL,
    maxNcharCatName = 10,
    dfMeta,
    group1ColorCol = "clusterColor", 
    group1NameCol = "clusterName",
    group2ColorCol = "sampleColor", 
    group2NameCol = "sampleName",
    columnSplitCol = "clusterName"
    ){
  
  if (is.null(chnkVec)){
    chnkVec <- as.vector(NULL, mode="character")
  }
  
  if (is.null(plotList)){
    plotList <- list()
  }
  
    allGenes <- unique(as.vector(unlist(geneGroupList)))
    
    
    if (is.null(cmdVec)){
      cmdVec <- as.vector(NULL, mode="character")
    }
    
    if (is.null(tag)){
      tag <- paste("Heatmap_", (length(plotList) + 1))
    }
    
    if (relativeExpression){
      legendString <- "log2 row mean"
    } else {
      legendString <- "lg10 Expr"
    }

# library("RColorBrewer")
# library("Seurat")
# library("ggplot2")
# library("viridis")
# #library("scMCA")
# library("circlize")
# library("ComplexHeatmap")



## Complex heatmap of cell/celltype correlations using just the top annotation per cell.  It's split by seurat cluster.
numbers_plot <- 1
# cors       <-  data.frame(OsC@assays$RNA$data)
#cors_index <- apply(cors, 2, gettissue, numbers_plot)
#cors_index <- sort(unique(as.integer(cors_index)))
allGenes <- allGenes[allGenes %in% row.names(dataMatrix)]

heat.dat   <- dataMatrix[allGenes, ]
#heat.dat <- heat.dat
#heat.dat   <- t(apply(heat.dat,1,function(x){(x-mean(x))/sd(x)}))

# Comment out for absolute expression

heat.dat <- heat.dat[apply(heat.dat, 1, mean) != 0, ]

if (relativeExpression){
  heat.dat   <- t(apply(heat.dat,1,function(x){log2((x/mean(x)))}))  
} 

heat.dat[heat.dat == -Inf] <- 0

## make sure cat names aren't too long
nMax <- max(nchar(names(geneGroupList)))

if (nMax > maxNcharCatName){
  shortNames <- sapply(
    names(geneGroupList),
    function(x){
      substr(x, 1, maxNcharCatName)
    }
  )
  
  if (sum(duplicated(shortNames)) > 0){
    indexVec <- 1:length(shortNames)
    shortNames[duplicated(shortNames)] <- paste0(shortNames[duplicated(shortNames)],"_", indexVec[duplicated(shortNames)])
  }
  
  names(geneGroupList) <- shortNames
  
}



rowSplitVec <- row.names(heat.dat)   
for (i in 1:length(geneGroupList)){
  rowSplitVec[rowSplitVec %in% geneGroupList[[i]]] <- names(geneGroupList)[i]
}



dfCol <- unique(dfMeta[,c(group1ColorCol, group1NameCol)])
colVec <- dfCol[,group1ColorCol]
names(colVec) <- dfCol[,group1NameCol]

dfCol2 <- unique(dfMeta[,c(group2ColorCol, group2NameCol)])
colVec2 <- dfCol2[,group2ColorCol]
names(colVec2) <- dfCol2[,group2NameCol]
# 
# colList <- list(
#   dfCol, 
#   dfCol2
# )


ht.anno <- ComplexHeatmap::HeatmapAnnotation(
    Sample =  dfMeta[,group2NameCol],
    Cluster =  dfMeta[,group1NameCol],
    col = list(
        Sample = colVec2,
        Cluster = colVec
    ),
    show_legend = showAllLegends
)


#bulkLogFC <- dfSel[allGenes, "contrast_2_logFC_p27NULL_E185OC_vs_WT_E185OC"]
#xcol <- ifelse(bulkLogFC < 0, "blue", "red")



#row_ha = rowAnnotation(Bulk = anno_barplot(bulkLogFC, gp = gpar(fill=xcol)),
#        show_legend = showAllLegends
#)   

#lrow_ha = rowAnnotation(foo = anno_block(gp = gpar(fill = c("blue", "red")),
#        labels = c("Down", "Up"), 
#        labels_gp = gpar(col = "white", fontsize = 10)),
#        show_legend = showAllLegends
#)
#     df = data.frame(
#         cluster = OsC_HM@meta.data$clusterName),
#     # col = list(
#     #     gp = gpar(col = "white")
#     # ),
#     annotation_legend_param = list(direction = "horizontal")
# )

if (relativeExpression){
  colRamp <- circlize::colorRamp2(c(-3, 0, 3),c("#3060cf", "#fffbbc","#c4463a"))
} else {
  colRamp <- circlize::colorRamp2(c(0, 6),c("#d3d3d3","#3060cf"))
}

if (nrow(heat.dat) > 100){
    rowFont <- 1
    show_row_names <- FALSE
} else if (nrow(heat.dat) > 50){
    rowFont <- 1
    show_row_names <- TRUE
} else if (nrow(heat.dat) > 25){
    rowFont <- 6
    show_row_names <- TRUE
} else {
    rowFont <- 8
    show_row_names <- TRUE
}

plotList[[tag]] <- ComplexHeatmap::Heatmap(
    heat.dat,
    row_split               = rowSplitVec,
    name                    = legendString,
    column_title_gp         = grid::gpar(fontsize = 6),
    row_title_rot           = 0, # Must be 0 or 90
    column_title_rot        = 90,
    column_split            = dfMeta[,columnSplitCol],
    #column_split            = column_split,
    cluster_column_slices   = cluster_column_slices,
    cluster_columns         = cluster_columns,
    cluster_rows            = TRUE,
    show_row_names          = show_row_names,
    show_column_names       = FALSE,
    column_names_side       = "bottom",
    show_column_dend        = TRUE,
    row_dend_width          = grid::unit(20, "mm"),
    show_heatmap_legend     = showAllLegends,
    column_names_max_height = grid::unit(8, "cm"),
    row_names_gp            = grid::gpar(fontsize = 6),
    top_annotation          = ht.anno,
    #right_annotation = row_ha,
    #left_annotation = lrow_ha,
    #col                     = colorRamp2(c(-2, 0, 2),magma(3)),
    col                     = colRamp,
    column_names_rot        = 90,
    border = TRUE
)

  FNbase <- paste0(tag, VersionPdfExt)

  FN <- paste0(Obio@parameterList$reportFigDir, FNbase)
  FNrel <- paste0("report_figures/", FNbase)
            
  pdf(FN)
    print(plotList[[tag]])
  dev.off()

  figLegend <- paste0(
    '**Figure ', 
    figureCount, 
    ':** ',
    'Heatmap split by cluster for gene set ',gsub("_", " ", tag),'. Download a pdf of this figure <a href="',FNrel,'" target="_blank">here</a>.'
  )
            
  figureCount <- figureCount + 1

  NewChnk <- paste0(
    "#### ", tag,
    "\n```{r hm_",
    tag,", results='asis', echo=F, eval=TRUE, warning=FALSE, fig.cap='",
    figLegend,"'}\n",
    "\n",
    "\n print(plotList[['",tag,"']])",
    "\n cat(  '\n')",
    "\n\n\n```\n"   
)

  chnkVec <- c(
    chnkVec,
    NewChnk
  )
  
  returnList <- list(
    "plotList" = plotList,
    "chnkVec"  = chnkVec,
    "figureCount" = figureCount
  )

return(returnList)
}

## Done function to create heatmaps
###############################################################################


#HMgeneSelections <- Obio@parameterList[["cat2HMplotList"]]

for (i in 1:length(HMgeneSelections)){
  for (j in 1:length(HMgeneSelections[[i]])){
    if (length(HMgeneSelections[[i]]) > 0){    
      tv <- HMgeneSelections[[i]][[j]]
      tv <- tv[tv %in% row.names(OsC)]
      ## Remove genes with all zero counts

      # tv <- tv[rowSums(OsC@assays$RNA[tv,]) != 0]
      HMgeneSelections[[i]][[j]] <- tv
    } 
  }  
}






# for (i in 1:length(HMgeneSelections)){
#   hl <- list(HMgeneSelections[[i]])
#   names(hl)[1] <- names(HMgeneSelections)[i]
#   geneList[[names(HMgeneSelections)[i]]]  <- hl
# }

#library(ComplexHeatmap)
plotList <- list()
chnkVec <- as.vector(NULL, mode="character")


## Make Heatmaps
if (!exists("figureCount")){
  figureCount <- 1  
}


## Relative Heatmaps for all gene sets
for (i in 1:length(HMgeneSelections)){

  if (length(unique(colnames(dataMatrix)) > 10)){
    showLegends <- FALSE
  } else {
    showLegends <- TRUE
  }
  
  if (length(unlist(HMgeneSelections[[i]])) > 2){
    returnList <- addHeatmap2List(
        dataMatrix = dataMatrix,
        cluster_column_slices   = FALSE,
        cluster_columns         = FALSE,
        geneGroupList = HMgeneSelections[[i]],
        relativeExpression = TRUE,
        cmdVec = NULL,
        tag = paste0(names(HMgeneSelections)[i], "_Rel"),
        showAllLegends = showLegends,
        figureCount = figureCount,
        plotList = plotList,
        maxNcharCatName = 10,
        dfMeta = dfMeta,
        group1ColorCol = "clusterColor", 
        group1NameCol = "clusterName",
        group2ColorCol = segmentationColorCol, 
        group2NameCol = segmentationCol,
        chnkVec = chnkVec,
        columnSplitCol = "clusterName"
        
    )
    
    plotList <- returnList[["plotList"]]
    chnkVec <- returnList[["chnkVec"]]
    figureCount <- returnList[["figureCount"]]
  }
}

## Absolute
for (i in 1:length(HMgeneSelections)){
  if (length(unlist(HMgeneSelections[[i]])) > 2){

    if (length(unique(colnames(dataMatrix)) > 10)){
        showLegends <- FALSE
    } else {
        showLegends <- TRUE
    }

    returnList <- addHeatmap2List(
        dataMatrix = dataMatrix,
        cluster_column_slices   = FALSE,
        cluster_columns         = FALSE,
        geneGroupList = HMgeneSelections[[i]],
        relativeExpression = FALSE,
        cmdVec = NULL,
        tag = paste0(names(HMgeneSelections)[i], "_Abs"),
        showAllLegends = showLegends,
        figureCount = figureCount,
        plotList = plotList,
        maxNcharCatName = 10,
        dfMeta = dfMeta,
        group1ColorCol = "clusterColor", 
        group1NameCol = "clusterName",
        group2ColorCol = segmentationColorCol, 
        group2NameCol = segmentationCol,
        chnkVec = chnkVec,
        columnSplitCol = "clusterName"
        
    )
    
    plotList <- returnList[["plotList"]]
    chnkVec <- returnList[["chnkVec"]]
    figureCount <- returnList[["figureCount"]]
  }
}





rm(OsC_HM)


## Done making Heatmap                                                       ##
###############################################################################


if (length(plotList) > 3){
    tabVar <- ".tabset .tabset-fade .tabset-dropdown"
} else {
    tabVar <- ".tabset .tabset-fade .tabset-pills"
}    
    
```


### Heatmaps {`r tabVar`}
:::: {.panel-tabset .nav-pills}
In this section heatmaps and dotplots for various gene categories are provided. 

```{r Cluster_overview, echo=TRUE, eval=TRUE, warning=FALSE, results="asis"}

cat(paste(knitr::knit(text = chnkVec, quiet = T), collapse = '\n'))

```
::::

```{r, eval=TRUE, echo=F, results=FALSE, message = FALSE, warning = FALSE}
tempDir <- "../../../../workdir/temp/"

if (!dir.exists(tempDir)){
  dir.create(tempDir, recursive = T)
}

if (exists("whiteListWorkspace")){
    rm(list = setdiff(ls(), whiteListWorkspace))
}

FN <- "../../../../workdir/temp/temp.workspace.RData"
save.image(FN)
```
